# backend/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API онлайн-сервиса для IT-школы.",
        "title": "skadi API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/solution/{id}": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение всех данных о решении задания с полной инфой о задании и преподе (ID и полное имя), а также со списком учеников, которые тоже выполняют это задание.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Получение решения задания по id [Преподаватель и ученик].",
                "operationId": "solution-read",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID решения задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.solutionOut"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "доступ запрещён"
                    },
                    "404": {
                        "description": "решение задания не найдено"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Удаление решения (не задания целиком) по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Удаление решения по id. [Только преподаватель]",
                "operationId": "solution-delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID решения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "доступ запрещён"
                    }
                }
            }
        },
...
        "/task": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение списка заданий конкретного преподавателя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Получение списка заданий [только преподаватель].",
                "operationId": "task-list",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "page pagination param",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "example": 5,
                        "description": "per page pagination param",
                        "name": "per-page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "F26",
                        "description": "substring to filter data by substring (case-insensitive)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.listTaskOut"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Создание нового задания для переданных учеников и учеников из переданных групп.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Создание нового задания. [Только преподаватель]",
                "operationId": "task-create",
                "parameters": [
                    {
                        "description": "taskBody",
                        "name": "taskBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.taskBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1.taskOut"
                        }
                    },
                    "400": {
                        "description": "неверный ученик | неверный преподаватель | преподаватель не найден"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    }
                }
            }
        },
        "/task/{id}": {
            "delete": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Удаление задания целиком (со всеми его решениями) по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Удаление задания по id. [Только преподаватель]",
                "operationId": "task-delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "доступ запрещён"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Частичное обновление задания (только переданные поля: название, описание, прикреплённые файлы) по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Обновление задания. [Только преподаватель]",
                "operationId": "task-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "updateTaskBody",
                        "name": "updateTaskBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.updateTaskBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Task"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "доступ запрещён"
                    },
                    "404": {
                        "description": "задание не найдено"
                    }
                }
            }
        },
        "/teacher/solution": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение списка решений для заданий конкретного преподавателя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Получение списка решений [только преподаватель].",
                "operationId": "teacher-solution-list",
                "parameters": [
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "filter for checked solutions if true",
                        "name": "archived",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "page pagination param",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "example": 5,
                        "description": "per page pagination param",
                        "name": "per-page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "F26",
                        "description": "substring to filter data by substring (case-insensitive)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.listSolutionOut"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    }
                }
            }
        },
        "/teacher/solution/{id}": {
            "patch": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Частичное обновление решения (только переданные поля: статус - \"готово\"/\"проверено\" - и оценка) по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Обновление решения. [Только преподаватель]",
                "operationId": "teacher-solution-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID решения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "updateSolutionBody",
                        "name": "updateSolutionBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.updateSolutionBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Solution"
                        }
                    },
                    "400": {
                        "description": "неверный статус"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "доступ запрещён"
                    },
                    "404": {
                        "description": "решение не найдено"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Profile": {
            "type": "object",
            "required": [
                "fullname"
            ],
            "properties": {
                "address": {
                    "description": "user address",
                    "type": "string"
                },
                "contact": {
                    "$ref": "#/definitions/entity.Contact"
                },
                "extra": {
                    "description": "extra data (for admin only)",
                    "type": "string"
                },
                "fullname": {
                    "description": "user full name",
                    "type": "string"
                },
                "id": {
                    "description": "profile id (equals to user id)",
                    "type": "integer"
                },
                "parent_contact": {
                    "$ref": "#/definitions/entity.Contact"
                }
            }
        },
        "entity.Solution": {
            "type": "object",
            "required": [
                "id",
                "status",
                "task"
            ],
            "properties": {
                "answer": {
                    "description": "solution text answer",
                    "type": "string"
                },
                "grade": {
                    "description": "solution grade",
                    "type": "string"
                },
                "id": {
                    "description": "solution id",
                    "type": "integer"
                },
                "status": {
                    "description": "status object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Status"
                        }
                    ]
                },
                "student": {
                    "description": "student object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                },
                "task": {
                    "description": "task object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Task"
                        }
                    ]
                },
                "updated_at": {
                    "description": "last-update datetime of solution",
                    "type": "string"
                }
            }
        },
...
        "entity.Task": {
            "type": "object",
            "required": [
                "id",
                "title"
            ],
            "properties": {
                "description": {
                    "description": "task description",
                    "type": "string"
                },
                "id": {
                    "description": "task id",
                    "type": "integer"
                },
                "teacher": {
                    "description": "teacher object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                },
                "title": {
                    "description": "task title",
                    "type": "string"
                }
            }
        },
...
        "v1.listSolutionOut": {
            "description": "listSolutionOut represents a solution list data.",
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "description": "solutions list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Solution"
                    }
                },
                "pagination": {
                    "description": "pagination params",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Pagination"
                        }
                    ]
                }
            }
        },
        "v1.listTaskOut": {
            "description": "listTaskOut represents a task list data.",
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "description": "tasks list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Task"
                    }
                },
                "pagination": {
                    "description": "pagination params",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Pagination"
                        }
                    ]
                }
            }
        },
...
        "v1.solutionOut": {
            "description": "solutionOut represents a solution data with students (solving the same task).",
            "type": "object",
            "required": [
                "solution"
            ],
            "properties": {
                "other_students": {
                    "description": "other students solving the same task",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Profile"
                    }
                },
                "solution": {
                    "description": "solution object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Solution"
                        }
                    ]
                }
            }
        },
        "v1.taskBody": {
            "description": "taskBody represents a data with task.",
            "type": "object",
            "required": [
                "description",
                "title"
            ],
            "properties": {
                "classes": {
                    "description": "classes for task solutions",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        3,
                        6,
                        9
                    ]
                },
                "description": {
                    "description": "task description",
                    "type": "string",
                    "example": "ООП в Python - это ..."
                },
                "students": {
                    "description": "students for task solutions",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        22,
                        32,
                        14
                    ]
                },
                "title": {
                    "description": "task title",
                    "type": "string",
                    "maxLength": 100,
                    "example": "ООП в Python"
                }
            }
        },
        "v1.taskOut": {
            "description": "taskOut represents a task data with solutions.",
            "type": "object",
            "required": [
                "task"
            ],
            "properties": {
                "solutions": {
                    "description": "task solutions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Solution"
                    }
                },
                "task": {
                    "description": "task object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Task"
                        }
                    ]
                }
            }
        },
...
        "v1.updateSolutionBody": {
            "description": "updateSolutionBody represents a data with optional body to update solution.",
            "type": "object",
            "properties": {
                "answer": {
                    "description": "new answer",
                    "type": "string",
                    "example": "ООП - это объектно-ориентированное программирование"
                },
                "grade": {
                    "description": "new grade",
                    "type": "string",
                    "maxLength": 50,
                    "example": "5+"
                },
                "status_id": {
                    "description": "new status ID",
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "v1.updateTaskBody": {
            "description": "updateTaskBody represents a data with optional body to update task.",
            "type": "object",
            "properties": {
                "description": {
                    "description": "new task description",
                    "type": "string",
                    "example": "Что такое ООП? Перечислить принципы ООП"
                },
                "title": {
                    "description": "new task title",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Понятие ООП"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTAccess": {
            "description": "Access JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "access",
            "in": "cookie"
        },
        "JWTRefresh": {
            "description": "Refresh JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "refresh",
            "in": "cookie"
        }
    }
}